refactor(csahier-v2): :art: move types under namespace, move config to the Router
