+++ packages/cashier-v2/src/types.ts
@@ -1,0 +1,19 @@
+import { cashierPathRoutes } from './routes/Router';
+
+// eslint-disable-next-line @typescript-eslint/no-namespace
+export namespace TRouteTypes {
+    export type TRoutes = typeof cashierPathRoutes[keyof typeof cashierPathRoutes];
+    export interface IRouteConfig {
+        component: React.ComponentType<IRouteConfig>;
+        path: string;
+        routes?: IRouteConfig[];
+        title: string;
+    }
+    export type TRouteComponent = React.ComponentProps<IRouteConfig['component']>;
+}
+
+declare module 'react-router-dom' {
+    export function useHistory(): { push: (path: TRouteTypes.TRoutes) => void };
+
+    export function useRouteMatch(path: TRouteTypes.TRoutes): boolean;
+}
